+++ me.cqp.luohuaming.BilibiliUpdateChecker.Code/OrderFunctions/AddStream.cs
@@ -37,12 +37,12 @@ namespace me.cqp.luohuaming.BilibiliUpdateChecker.Code.OrderFunctions
                 sendText.MsgToSend.Add("请填写用户ID");
                 return result;
             }
-            if (!int.TryParse(args, out int uid))
+            if (!long.TryParse(args, out long uid))
             {
                 sendText.MsgToSend.Add("用户ID格式不正确");
                 return result;
             }
-            var streams = JsonConfig.GetConfig<List<int>>("Streams");
+            var streams = JsonConfig.GetConfig<List<long>>("Streams");
             var group = JsonConfig.GetConfig<JObject>("Monitor_Stream");
             if (group.ContainsKey(e.FromGroup))
             {
